Generalise the IR metadata.
Instead of knowing of specific IR like `Span` or `StateIndex` it now has
a simple general representation of numbers, strings, references to other
metadata and structs.

Also, only a single metadatum reference may be attached to an IR value
or function at a time.  If multiple metadata are needed then they may be
aggregated in a `List` metadatum.